+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.0.3]
+
+### Changed
+- Internal Refactoring
+
+### Fixed
+- Fixed failure to patch static rpcs
+
 ## [v1.0.2]
 
 ### Fixed